Fix boss slime spawn

+++ core/src/com/upa/gun/Spawner.java
@@ -60,8 +60,8 @@ public class Spawner {
 
     void spawnBossSlime() {
         TextureRegion bossSlimeHurt = Assets.bossSlimeAnimations.get(ActionState.HURT).get(Direction.LEFT).getKeyFrame(0);
-        float spawnX = (Settings.RESOLUTION.x - bossSlimeHurt.getRegionWidth() / 2f / Settings.PPM);
-        float spawnY = (Settings.RESOLUTION.y - bossSlimeHurt.getRegionHeight() / 2f / Settings.PPM);
+        float spawnX = ((Settings.RESOLUTION.x - (float)bossSlimeHurt.getRegionWidth()) / 2f) / Settings.PPM;
+        float spawnY = ((Settings.RESOLUTION.y - (float)bossSlimeHurt.getRegionHeight()) / 2f) / Settings.PPM;
 
         BossSlimeFactory factory = new BossSlimeFactory();
         gunWorld.enemies.add(factory.makeBossSlime(bossHealth, spawnX, spawnY, world, gunWorld));
